compile on stable
This removes the use of the language features to enable its compilation
on rust stable.

the plugin feature was removed and so was the regex_macros dependency.
Instead of re-building the regular expression each time it needs to be
built, I build it once per `Context` instance, which I think is the next
best thing compared to a global, statically initialized regex.

the collections feature was needed for the str::slice_chars function. I
work around this for now by simply ripping out the slice_chars
implementation from the nightly std and embedding it in the code
directly, which we can use at least until the fate of std slice_chars
is determined.