Add compiletime feature to disable logging
If one uses a logger implementation which uses handlebars for "styling"
the output, logging calls in handlebars itself would start a recursion
and ultimatively crash the program by recursing indefinitly.

If the feature "logging" is not compiled into the library, the "log"
crate is ignored and the macros result in all logging calls to be
ignored.

By default, the "logging" feature is enabled to keep backwards
consistency.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>