Fix incorrect parse error for partial identifiers
This PR fixes a parse error that arises when using paths, containing
`/`, as partial identifiers.

The use case that brought this up was using
`Handlebars::register_templates_directory` which registers templates and
partials with the name set to the relative path from the root of the
templates directory.

```txt
- templates-dir/
  - bar.hbs
  - a/
    - foo.hbs
```

Here we have the partial `foo.hbs`. When we setup the `Handlebars`
registry with `Handlebars::register_templates_directory` we can then
refer to `foo.hbs` as `a/foo.hbs`.

The problem comes up with:

```hbs
{{! bar.hbs}}
{{#> content}}{{/content}}
```

```hbs
{{! foo.hbs}}
{{#> a/bar}}
  {{#*inline "content"}}SOME CONTENT{{/inline}}
{{/a/bar}}
```

Specifically, the parse error shows up on the closing expression of the
partial block (the line `{{/a/bar}}` in `foo.hbs`.