expose DirectorySourceOptionsBuilder & added doc example
DirectorySourceOptions is `#[non_exhaustive]` which means it cannot be
created directly by users of the crate.

DirectorySourceOptions is `#[derive(Builder)]` but the resulting
DirectorySourceOptionsBuilder is not exposed to users of the crate.

Because of this only `DirectorySourceOptions::default()` can be used and
options cannot be set.

Exposing DirectorySourceOptionsBuilder allows users of this crate to
build DirectorySourceOptions.